Implemented simple search algorithm to attempt to get module working
Although this algorithm probably isn't the fastest, it should work once I finish it.  It won't work, though, until the function to turn the image black and white and the function to read a pixel value from the image work.  cv::floodFill will likely be a better solution in the future, though.